feat: add homepage view refinement
